Add client home component and middleware fallback
